레이아웃 및 메인 페이지 개선: 사이드바 추가, 메타데이터 수정, 한국어 지원, 유사 상품 탐색 및 판매 지표 예측 섹션 추가
