+++ app/layout.tsx
@@ -1,10 +1,13 @@
 import type { Metadata } from 'next'
+import { Inter } from 'next/font/google'
 import './globals.css'
+import Sidebar from '@/components/Sidebar'
+
+const inter = Inter({ subsets: ['latin'] })
 
 export const metadata: Metadata = {
-  title: 'v0 App',
-  description: 'Created with v0',
-  generator: 'v0.dev',
+  title: 'Rmazon - Amazon Market Analysis',
+  description: 'Amazon 시장에서 유사한 제품을 찾아 경쟁력을 분석하고 판매 지표를 예측해보세요.',
 }
 
 export default function RootLayout({
@@ -13,8 +16,15 @@ export default function RootLayout({
   children: React.ReactNode
 }>) {
   return (
-    <html lang="en">
-      <body>{children}</body>
+    <html lang="ko">
+      <body className={inter.className}>
+        <div className="flex h-screen bg-gray-50/50">
+          <Sidebar />
+          <main className="flex-1 overflow-y-auto">
+            {children}
+          </main>
+        </div>
+      </body>
     </html>
   )
 }
